@@ -1,33 +1,64 @@
 {
-   "name": "qrcode-service",
-   "version": "1.0.0",
-   "main": "index.js",
-   "license": "MIT",
-   "scripts": {
-      "dev": "ts-node-dev --respawn --transpileOnly src/index.ts",
-      "start": "ts-node src/index.ts",
-      "typeorm":"ts-node ./node_modules/typeorm/cli.js"
-   },
-   "dependencies": {
-      "@marblejs/core": "^3.1.2",
-      "@marblejs/middleware-body": "^3.1.2",
-      "@marblejs/middleware-io": "^3.1.2",
-      "@marblejs/middleware-jwt": "^3.1.2",
-      "@marblejs/middleware-logger": "^3.1.2",
-      "dotenv": "^8.2.0",
-      "fp-ts": "^2.6.2",
-      "mysql": "^2.14.1",
-      "pg": "^8.2.1",
-      "qrcode": "^1.4.4",
-      "reflect-metadata": "^0.1.10",
-      "rxjs": "^6.5.5",
-      "typeorm": "0.2.25"
-   },
-   "devDependencies": {
-      "@types/node": "^8.0.29",
-      "@types/qrcode": "^1.3.4",
-      "ts-node": "3.3.0",
-      "ts-node-dev": "^1.0.0-pre.44",
-      "typescript": "3.3.3333"
-   }
+  "name": "qrcode-service",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "scripts": {
+    "clear": "rimraf dist",
+    "start": "node dist/index.js",
+    "build": "yarn clear && tsc && cp ormconfig.prod.js ./dist",
+    "dev": "ts-node-dev --respawn --transpileOnly src/index.ts",
+    "typeorm": "ts-node ./node_modules/typeorm/cli.js --config ormconfig.prod.js",
+    "typeorm:prod": "ts-node-dev ./node_modules/typeorm/cli.js --config dist/ormconfig.prod.js",
+    "typeorm:dev": "ts-node-dev ./node_modules/typeorm/cli.js --config ormconfig.dev.js"
+  },
+  "dependencies": {
+    "@marblejs/core": "^3.1.2",
+    "@marblejs/middleware-body": "^3.1.2",
+    "@marblejs/middleware-io": "^3.1.2",
+    "@marblejs/middleware-jwt": "^3.1.2",
+    "@marblejs/middleware-logger": "^3.1.2",
+    "add": "^2.0.6",
+    "core-js": "^3.6.5",
+    "dotenv": "^8.2.0",
+    "fp-ts": "^2.6.2",
+    "mysql": "^2.14.1",
+    "pg": "^8.2.1",
+    "qrcode": "^1.4.4",
+    "reflect-metadata": "^0.1.10",
+    "rxjs": "^6.5.5",
+    "tsconfig-paths": "^3.9.0",
+    "typeorm": "0.2.25"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.10.1",
+    "@babel/core": "^7.10.2",
+    "@babel/node": "^7.10.1",
+    "@babel/plugin-transform-typescript": "^7.10.1",
+    "@babel/preset-env": "^7.10.2",
+    "@types/dotenv": "^8.2.0",
+    "@types/node": "^8.0.29",
+    "@types/qrcode": "^1.3.4",
+    "@types/webpack-node-externals": "^1.7.1",
+    "@typescript-eslint/eslint-plugin": "^3.0.2",
+    "@typescript-eslint/parser": "^3.0.2",
+    "awesome-typescript-loader": "^5.2.1",
+    "babel-polyfill": "^6.26.0",
+    "eslint": "^7.1.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-promise": "^4.2.1",
+    "prettier": "^2.0.5",
+    "rimraf": "^3.0.2",
+    "source-map-loader": "^1.0.0",
+    "ts-loader": "^7.0.5",
+    "ts-node": "3.3.0",
+    "ts-node-dev": "^1.0.0-pre.44",
+    "typescript": "^3.9.3",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-node-externals": "^1.7.2"
+  }
 }
