@@ -1,19 +1,24 @@
-import { HttpEffect, HttpError, HttpStatus } from "@marblejs/core";
-import { mergeMap, map, catchError } from "rxjs/operators";
-import { of, throwError } from "rxjs";
-import { create } from "../../models/QRCode/qr-code.models";
+import { HttpEffect, HttpError, HttpStatus } from '@marblejs/core';
+import { mergeMap, map, catchError } from 'rxjs/operators';
+import { of, throwError } from 'rxjs';
+import { create } from '../../models/QRCode/qr-code.models';
+import { generateQRCode } from '../../models/QRCode';
 
-export const qrCodeEffects$: HttpEffect = (req$) =>
+export const qrCodeEffects$: HttpEffect = req$ =>
   req$.pipe(
-    mergeMap((req) =>
+    mergeMap(req =>
       of(req).pipe(
-        map((req) => req.body, console.log(req.body)),
+        // map((req) => req.body, console.log(req.query)),
+        map(req => req.body),
         mergeMap(create),
-        map(({ code }) => ({
+        mergeMap(res =>
+          generateQRCode(String(res.code), Number(res.scale || 4))
+        ),
+        map(body => ({
           status: HttpStatus.CREATED,
-          body: { code },
+          body,
         })),
-        catchError((err) =>
+        catchError(err =>
           throwError(new HttpError(err, HttpStatus.BAD_REQUEST))
         )
       )
