@@ -1,37 +1,42 @@
-import { MigrationInterface, QueryRunner, Table } from "typeorm";
+import { MigrationInterface, QueryRunner, Table } from 'typeorm';
 
 export class qrcode1590706465433 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query('CREATE EXTENSION IF NOT EXISTS "uuid-ossp";');
     await queryRunner.createTable(
       new Table({
-        name: "qr_code",
+        name: 'qr_code',
         columns: [
           {
-            name: "id",
-            type: "uuid",
-            default: "uuid_generate_v4()",
+            name: 'id',
+            type: 'uuid',
+            default: 'uuid_generate_v4()',
             isPrimary: true,
           },
           {
-            name: "code",
-            type: "varchar",
+            name: 'code',
+            type: 'varchar',
             isNullable: false,
           },
           {
-            name: "expire_at",
-            type: "date",
+            name: 'expire_at',
+            type: 'timestamp',
             isNullable: false,
           },
           {
-            name: "created_at",
-            type: "timestamptz",
-            default: "CURRENT_TIMESTAMP",
+            name: 'scale',
+            type: 'integer',
+            default: 4,
           },
           {
-            name: "updated_at",
-            type: "timestamptz",
-            default: "CURRENT_TIMESTAMP",
+            name: 'created_at',
+            type: 'timestamptz',
+            default: 'CURRENT_TIMESTAMP',
+          },
+          {
+            name: 'updated_at',
+            type: 'timestamptz',
+            default: 'CURRENT_TIMESTAMP',
           },
         ],
       }),
@@ -40,6 +45,6 @@ export class qrcode1590706465433 implements MigrationInterface {
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.dropTable("qr_code");
+    await queryRunner.dropTable('qr_code');
   }
 }
